requirements.txt
openai==0.28.*
langchain>=0.0.188
tenacity>=8.2.2
tiktoken>=0.4.0
rich>=13.0.1
chromadb>=0.3.25
pdf2image>=1.16.3
python-dotenv>=1.0.0
ipykernel>=6.23.3
arxiv==1.4.8
streamlit==1.27.*
latex2markdown
